refactor: split _model_helpers.py into 3 focused mixin files
The 625-line _model_helpers.py mixed three concerns under one
_ModelHelpersMixin. Split into:

  _intervals.py (256 lines) — _IntervalsMixin
    bindings + break-time + locked/unavailable interval creation.
  _dependencies.py (308 lines) — _DependenciesMixin
    travel-time encodings (direct / pairwise / table / hybrid) +
    _machine_choices + _table_machine_task_ids.
  _cumulatives.py (130 lines) — _CumulativesMixin
    per-machine disjunctive (no_overlap or cumulative) + redundant
    per-capability cumulative.

CpSatSolver inherits the three new mixins (alongside the existing
_TaskBuildMixin / _HeuristicMixin / _ObjectiveMixin). The
self.method() call sites are unchanged because every method lives on
`self` via MRO. All 140 tests pass.
